+++ src/main/java/net/abraxator/moresnifferflowers/client/gui/screen/GluedOverlay.java
@@ -1,0 +1,24 @@
+package net.abraxator.moresnifferflowers.client.gui.screen;
+
+import net.abraxator.moresnifferflowers.MoreSnifferFlowers;
+import net.abraxator.moresnifferflowers.init.ModEffects;
+import net.minecraft.client.Minecraft;
+import net.minecraft.client.gui.GuiGraphics;
+import net.minecraft.client.player.LocalPlayer;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraftforge.client.gui.overlay.ForgeGui;
+import net.minecraftforge.client.gui.overlay.IGuiOverlay;
+
+public class GluedOverlay implements IGuiOverlay {
+    @Override
+    public void render(ForgeGui gui, GuiGraphics guiGraphics, float partialTick, int width, int height) {
+        Minecraft minecraft = Minecraft.getInstance();
+        LocalPlayer player = minecraft.player;
+
+        if (player == null || !player.hasEffect(ModEffects.GLUED.get())) return;
+
+        ResourceLocation texture = MoreSnifferFlowers.loc("textures/gui/glued_overlay.png");
+        guiGraphics.blit(texture, 0, 0, 0, 0, width, height, width, height);
+
+    }
+}