+++ CHANGELOG.md
@@ -1,2 +1,1 @@
-- Converted the bobling music disc from stereo to mono so minecraft stops complaining
-- Gotten rid of a few startup errors
+- Fixed performance issues when rendering the glued overlay